Uninstall - Remove Video Captures script - check to make sure it exists before trying to delete it

+++ InstallScripts/Utility/PowerShell/RemoveVideoCaptures.ps1
@@ -5,7 +5,10 @@
     $videoCapturesRoot = "C:\VideoCaptures"
 
     Write-Host "Removing Video Captures..." -NoNewline
-    Remove-Item $videoCapturesRoot -recurse -Force
+    If(test-path $videoCapturesRoot)
+    {
+        Remove-Item $videoCapturesRoot -recurse -Force
+    }
     Write-Host "done."
 }
 Catch
